+++ galaxy/quasar/src/qs/graphics/Particle.cpp
@@ -13,12 +13,12 @@
 namespace qs
 {
 	Particle::Particle()
-	    : m_life {100.0f}, m_position {0.0f, 0.0f}, m_velocity {1.0f, 1.0f}
+	    : m_life {1.0f}, m_position {0.0f, 0.0f}, m_velocity {1.0f, 1.0f}
 	{
 	}
 
 	Particle::Particle(const float x_vel, const float y_vel)
-	    : m_life {100.0f}, m_position {0.0f, 0.0f}, m_velocity {x_vel, y_vel}
+	    : m_life {1.0f}, m_position {0.0f, 0.0f}, m_velocity {x_vel, y_vel}
 	{
 	}
 
@@ -83,43 +83,5 @@ namespace qs
 	const glm::vec2& Particle::velocity() const noexcept
 	{
 		return m_velocity;
-	}
-
-	ParticleInstance::ParticleInstance(ParticleInstance&& p)
-	{
-		this->m_instance_buffer = std::move(p.m_instance_buffer);
-	}
-
-	ParticleInstance& ParticleInstance::operator=(ParticleInstance&& p)
-	{
-		if (this != &p)
-		{
-			this->m_instance_buffer = std::move(p.m_instance_buffer);
-		}
-
-		return *this;
-	}
-
-	void ParticleInstance::set_instance(std::span<glm::vec2> instances)
-	{
-		m_instance_buffer.create(instances, 1);
-		m_va.set_instanced(m_instance_buffer);
-	}
-
-	void ParticleInstance::update_instances(std::span<glm::vec2> instances)
-	{
-		m_instance_buffer.update(instances);
-	}
-
-	void ParticleInstance::bind() noexcept
-	{
-		m_va.bind();
-		glBindTexture(GL_TEXTURE_2D, m_texture);
-	}
-
-	void ParticleInstance::unbind() noexcept
-	{
-		m_va.unbind();
-		glBindTexture(GL_TEXTURE_2D, 0);
 	}
 } // namespace qs
