+++ sandbox_tests/CMakeLists.txt
@@ -135,3 +135,12 @@ elseif (GNU)
         -static
     )
 endif()
+
+# Postbuild shader copy.
+if("${CMAKE_BUILD_TYPE}" STREQUAL "Debug")
+    add_custom_command(TARGET sandbox_test POST_BUILD COMMAND ${CMAKE_COMMAND} -E copy_directory "${CMAKE_SOURCE_DIR}/galaxy/quasar/shaders" "${CMAKE_SOURCE_DIR}/sandbox_tests/bin/Debug/assets/shaders")
+elseif("${CMAKE_BUILD_TYPE}" STREQUAL "Release")
+    add_custom_command(TARGET sandbox_test POST_BUILD COMMAND ${CMAKE_COMMAND} -E copy_directory "${CMAKE_SOURCE_DIR}/galaxy/quasar/shaders" "${CMAKE_SOURCE_DIR}/sandbox_tests/bin/Debug/assets/shaders")
+else()
+    message(FATAL_ERROR "Could not determine build configuration. Is currently: ${CMAKE_BUILD_TYPE}")
+endif()
